requirements.txt
addict==2.4.0
albumentations==1.3.1
braceexpand==0.1.7
diffusers==0.26.3
easydict==1.11
gradio==4.19.0
huggingface_hub==0.20.2
imageio==2.31.6
imgaug==0.4.0
ipdb==0.13.13
joblib==1.3.2
kornia==0.6.7
matplotlib==3.8.0
memory_profiler==0.61.0
numpy
omegaconf==2.2.3
onnxruntime==1.15.0
opencv_contrib_python==4.8.1.78
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
pandas==2.2.0
#Pillow==9.5.0
Pillow==10.2.0
pycocotools==2.0.7
pytorch_lightning==1.9.4
PyYAML==6.0.1
PyYAML==6.0.1
Requests==2.31.0
scikit_learn==1.3.2
scipy==1.12.0
setuptools==60.2.0
six==1.16.0
scikit-image
streamlit==1.31.1
supervision==0.18.0
sync_batchnorm==0.0.1
tensorflow==2.15.0.post1
termcolor==2.4.0
timm==0.4.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.65.0
transformers==4.25.1
webdataset==0.2.86
yapf==0.40.2
